Use current Node.js supported releases
Node.js 8.x left maintenance at the end of 2019, and 10.x left
maintenance at the end of April. 12.x is planned to be maintained until
2022-04-30. 14.x is the current active LTS, and is planned to be
maintained until 2023-04-30, and 16.x is the current release, planned to
be maintained until 2024-04-30.

Drop the end-of-life versions from the test matrix, and add the
currently supported versions.

See also:
- https://en.wikipedia.org/wiki/Node.js#Releases
- https://github.com/nodejs/Release